Exercise 7.13 complete
