+++ bloglist-frontend/src/components/Login.jsx
@@ -1,0 +1,48 @@
+import { useState } from "react";
+import { useDispatch } from "react-redux";
+import { userLogin } from "../reducers/userReducer";
+
+const Login = () => {
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+
+  const dispatch = useDispatch();
+
+  const handleLogin = async (e) => {
+    e.preventDefault();
+    dispatch(userLogin(username, password));
+    setUsername("");
+    setPassword("");
+  };
+
+  return (
+    <>
+      <h2>Log in to application</h2>
+      <form onSubmit={handleLogin}>
+        <div>
+          username
+          <input
+            id="username"
+            type="text"
+            value={username}
+            name="Username"
+            onChange={({ target }) => setUsername(target.value)}
+          ></input>
+        </div>
+        <div>
+          password
+          <input
+            id="password"
+            type="password"
+            value={password}
+            name="Password"
+            onChange={({ target }) => setPassword(target.value)}
+          ></input>
+        </div>
+        <button type="submit">login</button>
+      </form>
+    </>
+  );
+};
+
+export default Login;
